Split offscreen graph setup
